+++ pymc3/distributions/continuous.py
@@ -523,7 +523,6 @@ class TruncatedNormal(BoundedContinuous):
         a1 = [-3, -5, -5]
         b1 = [7, 5, 4]
         for mu, sigma, a, b in zip(mus, sigmas,a1,b1):
-            print mu, sigma, a, b
             an, bn = (a - mu) / sigma, (b - mu) / sigma
             pdf = st.truncnorm.pdf(x, an,bn, loc=mu, scale=sigma)
             plt.plot(x, pdf, label=r'$\mu$ = {}, $\sigma$ = {}, a={}, b={}'.format(mu, sigma, a, b))
@@ -2366,6 +2365,18 @@ class Gamma(PositiveContinuous):
         return bound(
             -gammaln(alpha) + logpow(
                 beta, alpha) - beta * value + logpow(value, alpha - 1),
+            value >= 0,
+            alpha > 0,
+            beta > 0)
+
+    def logcdf(self, value):
+        """
+        Compute the log CDF for the Gamma distribution
+        """
+        alpha = self.alpha
+        beta = self.beta
+        return bound(
+            tt.log(tt.gammainc(alpha, beta * value)),
             value >= 0,
             alpha > 0,
             beta > 0)